ckean_up_fixes
