+++ src/start_menu/main_settings/ChangeFontWeight.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void Weight();//Declares the Weight() function to prompt the user for a weight choice and customize text weight
