+++ src/start_menu/main_settings/font.h
@@ -1,4 +1,3 @@
 #pragma once
-#include "../start_menu.h"
 
-void Font();
+void Font(); // Declares the Font() function to prompt the user to select the font
