+++ src/start_menu/main_settings/main_settings.h
@@ -1,4 +1,3 @@
 #pragma once
-#include "../start_menu.h"
 
-void MainSettings();
+void MainSettings(); // Declares a list of available settings options and prompts the user to select one
