+++ src/start_menu/main_settings/size.cpp
@@ -2,8 +2,9 @@
 #include "../../utils/utils.h"
 #include <iostream>
 
-void Size(){
-	std::cout << "Pick Number" << std::endl;
+void Size()
+{
+	std::cout << "Pick text size number" << std::endl;
 	utils::GetUserInput(utils::g_text_size);
 	utils::CustomizeText(utils::g_text_size);
 }
