+++ src/start_menu/main_settings/speed.cpp
@@ -3,7 +3,6 @@
 #include <iostream>
 
 void Speed(){
-	std::cout << "Give me the time you want between each rendered "
-				 "character (in ms): ";
+	std::cout << "Give me the time you want between each rendered character (in ms): ";
 	utils::GetUserInput(utils::g_sleep_for_ms);
 }
