--- src/start_menu/main_settings/weight.h
@@ -1,5 +1,0 @@
-#pragma once
-#include "../start_menu.h"
-
-
-void Weight();
